+++ src/initapp.php
@@ -37,3 +37,17 @@ $container['logger'] = function ($c) {
 	$logger->pushHandler($file_handler);
 	return $logger;
 };
+
+
+// Register component on container
+$container['view'] = function ($container) {
+    $view = new \Slim\Views\Twig(__DIR__ .  '/ui', [
+        'cache' => false
+    ]);
+
+    // Instantiate and add Slim specific extension
+    $basePath = rtrim(str_ireplace('index.php', '', $container->get('request')->getUri()->getBasePath()), '/');
+    $view->addExtension(new \Slim\Views\TwigExtension($container->get('router'), $basePath));
+
+    return $view;
+};