+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## dev
+## dev (0.1.0)
 - delete a single job
 - `all_jobs` shows all jobs in the `jobs/status` route.
+- stop server using `.stop_server` file or `/server/stop` route
 
 ## [0.0.2] - 2018-08-06
 - `__JOBID__` Placeholder